Improvements for cross-platform ANGLing
MinGW from Linux is a good test case for getting all the OpenGL parts
working. This starts the process of getting that functional.

We may need to start storing versioned copies of generated headers in
order to support this use case. While it's probably not impossible to
generate FXC headers on Linux, it's much easier to store the generated
files somewhere. ANGLE does this upstream already anyway.